converted to use aspnet health check
